backend/requirements.txt
fastapi
uvicorn
transformers
optimum[onnxruntime]
chromadb
sentence-transformers
google-genai
python-dotenv
pydantic
